fixing devlop02 in my repo not the template
